[Update] 스크롤뷰 설정 수정
Content size fitter component 추가하여 정상적으로 스크롤하도록 수정

+++ Assets/Scripts/Managers/AchievementDataManager.cs
@@ -239,19 +239,19 @@ public class AchievementDataManager : SingletonBase<AchievementDataManager>, ISa
         {
             case AchievementType.HighScore:
             {
-                progress = PlayDataManager.Instance.GetHighscore() / requiredvalue;
+                progress = PlayDataManager.Instance.GetHighscore() / (float)requiredvalue;
                 if (progress >= 1) return 1;
                 else return progress;
             }
             case AchievementType.TotalCoin:
             {
-                progress = PlayDataManager.Instance.GetTotalCoin() / requiredvalue;
+                progress = PlayDataManager.Instance.GetTotalCoin() / (float)requiredvalue;
                 if (progress >= 1) return 1;
                 else return progress;
             }
             case AchievementType.UnlockCharacters:
             {
-                progress = PlayDataManager.Instance.GetUnlockedCharacterCount() / requiredvalue;
+                progress = PlayDataManager.Instance.GetUnlockedCharacterCount() / (float)requiredvalue;
                 if (progress >= 1) return 1;
                 else return progress;
             }
